+++ src/components/world/SceneView.tsx
@@ -696,6 +696,7 @@ const SceneView = memo(
                 width={scene.width}
                 height={scene.height}
                 collisions={scene.collisions}
+                sceneTypeKey={scene.type}
               />
               {selected && slopePreview && (
                 <SceneSlopePreview